Fix data race with gRPC logging during program termination (#2353)
