+++ cloud/blockstore/libs/client/client.cpp
@@ -26,7 +26,7 @@
 #include <cloud/storage/core/libs/grpc/channel_arguments.h>
 #include <cloud/storage/core/libs/grpc/credentials.h>
 #include <cloud/storage/core/libs/grpc/executor.h>
-#include <cloud/storage/core/libs/grpc/initializer.h>
+#include <cloud/storage/core/libs/grpc/init.h>
 #include <cloud/storage/core/libs/grpc/time_point_specialization.h>
 #include <cloud/storage/core/libs/uds/uds_socket_client.h>
