+++ cloud/blockstore/libs/discovery/ping.cpp
@@ -3,7 +3,7 @@
 #include <cloud/blockstore/public/api/grpc/service.grpc.pb.h>
 
 #include <cloud/storage/core/libs/common/error.h>
-#include <cloud/storage/core/libs/grpc/initializer.h>
+#include <cloud/storage/core/libs/grpc/init.h>
 #include <cloud/storage/core/libs/grpc/time_point_specialization.h>
 
 #include <contrib/libs/grpc/include/grpcpp/channel.h>
