+++ cloud/storage/core/libs/grpc/init.cpp
@@ -1,10 +1,12 @@
-#include "logging.h"
+#include "init.h"
 
 #include <contrib/libs/grpc/include/grpc/grpc.h>
 #include <contrib/libs/grpc/include/grpc/support/log.h>
 
+#include <library/cpp/deprecated/atomic/atomic.h>
 #include <library/cpp/logger/log.h>
 
+#include <util/generic/singleton.h>
 #include <util/system/src_location.h>
 
 #include <atomic>
@@ -14,6 +16,29 @@ namespace NCloud {
 namespace {
 
 ////////////////////////////////////////////////////////////////////////////////
+
+struct TGrpcState
+{
+    TAtomic Counter = 0;
+    TLog Log;
+
+    ~TGrpcState()
+    {
+        // ensure that GRPC is shut down
+        Y_ABORT_UNLESS(AtomicGet(Counter) == 0);
+    }
+
+    static TGrpcState& Instance()
+    {
+        constexpr auto defaultPriority =
+            TSingletonTraits<TGrpcState>::Priority;
+        static_assert(defaultPriority > 0);
+
+        // need priority less than default priority to ensure that TGrpcState
+        // singleton lives longer than other singletons (like TApp)
+        return *SingletonWithPriority<TGrpcState, defaultPriority - 1>();
+    }
+};
 
 gpr_log_severity LogPriorityToSeverity(ELogPriority priority)
 {
@@ -39,15 +64,15 @@ ELogPriority LogSeverityToPriority(gpr_log_severity severity)
     }
 }
 
-TLog Log;
-
 void AddLog(gpr_log_func_args* args)
 {
     auto file = ::NPrivate::StripRoot({
         args->file,
         static_cast<ui32>(strlen(args->file))});
 
-    Log << LogSeverityToPriority(args->severity)
+    auto& state = TGrpcState::Instance();
+
+    state.Log << LogSeverityToPriority(args->severity)
         << TSourceLocation(file.As<TStringBuf>(), args->line)
         << ": " << args->message;
 }
@@ -77,9 +102,30 @@ void EnableGRpcTracing()
 
 ////////////////////////////////////////////////////////////////////////////////
 
+TGrpcInitializer::TGrpcInitializer()
+{
+    auto& state = TGrpcState::Instance();
+
+    if (AtomicGetAndIncrement(state.Counter) == 0) {
+        grpc_init();
+    }
+}
+
+TGrpcInitializer::~TGrpcInitializer()
+{
+    auto& state = TGrpcState::Instance();
+
+    if (AtomicDecrement(state.Counter) == 0) {
+        grpc_shutdown_blocking();
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
 void GrpcLoggerInit(const TLog& log, bool enableTracing)
 {
-    Log = log;
+    auto& state = TGrpcState::Instance();
+    state.Log = log;
     // |gpr_set_log_verbosity| and |gpr_set_log_function| do not imply any
     // memory barrier, so we need a full barrier here.
     std::atomic_thread_fence(std::memory_order_seq_cst);
